Refactor: Rename wdctl_thread to wdctlx_thread
I've renamed src/wdctl_thread.c to src/wdctlx_thread.c and src/wdctl_thread.h to src/wdctlx_thread.h.

I also updated all corresponding references within the codebase, including the #include directive in wdctlx_thread.c and entries in src/CMakeLists.txt.

I attempted to verify the build but couldn't complete it due to an issue with my environment.
